chore(firefox): raise the minimum supported version to 128
Firefox 128 is the first version where content scripts can target
execution worlds natively (world and matchOriginAsFallback). Requiring
it removes the page-bootstrap fallback for MAIN-world scriptlets and
the dedicated minimum-Firefox test infrastructure.
