+++ .github/workflows/test.yml
@@ -86,25 +86,3 @@ jobs:
         run: npm ci
       - name: Run tests after updating the extension
         run: npm run wdio:update -- --target=firefox
-
-  # Scriptlet injection on the minimum supported Firefox (115), where the `world`
-  # and `matchOriginAsFallback` content-script options are not available.
-  e2e-firefox-min:
-    runs-on: ubuntu-latest
-    needs: unit
-    steps:
-      - uses: actions/checkout@v5
-      - uses: actions/setup-node@v5
-        with:
-          node-version-file: '.tool-versions'
-      # Firefox < 121 cannot start without libdbus-glib, which Ubuntu 24.04 no longer ships.
-      - name: Install Firefox host dependencies
-        run: sudo apt-get update && sudo apt-get install -y --no-install-recommends libdbus-glib-1-2
-      # Firefox 115 content processes segfault when AppArmor (Ubuntu 24.04) denies
-      # unprivileged user namespaces; newer Firefox falls back gracefully.
-      - name: Allow Firefox sandbox user namespaces
-        run: sudo sysctl -w kernel.apparmor_restrict_unprivileged_userns=0
-      - name: Install dependencies
-        run: npm ci
-      - name: Run scriptlet execution-world tests
-        run: npm run wdio -- --target=firefox --firefox-version=esr_115.0 --logLevel=warn --spec tests/e2e/spec/adblocker-world.spec.js
